+++ modules/m_pong.c
@@ -104,7 +104,7 @@ mr_pong(struct MsgBuf *msgbuf_p, struct Client *client_p, struct Client *source_
 	{
 		if(ConfigFileEntry.ping_cookie && source_p->flags & FLAGS_SENTUSER && source_p->name[0])
 		{
-			unsigned long incoming_ping = strtoul(parv[1], NULL, 16);
+			uint32_t incoming_ping = strtoul(parv[1], NULL, 16);
 			if(incoming_ping)
 			{
 				if(source_p->localClient->random_ping == incoming_ping)
